refactor(housekeeping): move migrator to API module
mostly a copy-paste of files, the big change was starting `:ecto` and
data access in the release task

code from the backend module has been deleted as it was not used anymore

a single test file was kept in order to allow coverage reports and not
fail in ci
