--- apps/astarte_housekeeping/entrypoint.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-./bin/astarte_housekeeping eval Elixir.Astarte.Housekeeping.ReleaseTasks.init_database || exit 1
-
-./bin/astarte_housekeeping eval Elixir.Astarte.Housekeeping.ReleaseTasks.migrate || exit 1
-
-exec ./bin/astarte_housekeeping $@
