+++ apps/astarte_housekeeping/test/test_helper.exs
@@ -1,6 +1,3 @@
 Mimic.copy(Xandra)
 Mimic.copy(Xandra.Cluster)
-Mimic.copy(Astarte.Housekeeping.Queries)
 ExUnit.start(capture_log: true)
-Mimic.copy(Astarte.Housekeeping.Queries)
-ExUnit.start()
